refactor: changes to suppress problems of first migration

+++ config/sentry.php
@@ -14,5 +14,5 @@ return [
 
     'send_default_pii' => env('SENTRY_COLLECT_USERS', false),
 
-    'traces_sample_rate' => env('SENTRY_TRACES_SIMPLE_RATE', 0.0),
+    'traces_sample_rate' => (float) env('SENTRY_TRACES_SIMPLE_RATE', 0.0),
 ];
